+++ src/converter/index.ts
@@ -6,12 +6,15 @@ import {
   DuoAccountSettings,
   DuoGroup,
   DuoAdmin,
+  DuoPhone,
+  DuoIntegration,
 } from '../collector/types';
 import {
   createIntegrationEntity,
   getTime,
   convertProperties,
 } from '@jupiterone/integration-sdk-core';
+import { Entities } from '../constants';
 
 export function convertAccount(
   siteId: string,
@@ -25,8 +28,8 @@ export function convertAccount(
         _key: `duo-account:${siteId}:${data.name
           .toLowerCase()
           .replace(/\s/, '-')}`,
-        _type: 'duo_account',
-        _class: 'Account',
+        _type: Entities.ACCOUNT._type,
+        _class: Entities.ACCOUNT._class,
         name: data.name,
         siteId,
         displayName: `${data.name} Duo Account`,
@@ -44,8 +47,8 @@ export function convertUser(
       source: user,
       assign: {
         _key: `duo-user:${user.user_id}`,
-        _type: 'duo_user',
-        _class: 'User',
+        _type: Entities.USER._type,
+        _class: Entities.USER._class,
         id: user.user_id,
         name: user.realname,
         firstName: user.firstname,
@@ -73,8 +76,8 @@ export function convertGroup(
       assign: {
         ...convertProperties(group),
         _key: `duo-group:${group.group_id}`,
-        _type: 'duo_group',
-        _class: 'UserGroup',
+        _type: Entities.USER_GROUP._type,
+        _class: Entities.USER_GROUP._class,
         id: group.group_id,
         description: group.desc,
         active: group.status.toLowerCase() === 'active',
@@ -92,8 +95,8 @@ export function convertAdmin(
       assign: {
         ...convertProperties(admin),
         _key: `duo-admin:${admin.admin_id}`,
-        _type: 'duo_admin',
-        _class: 'User',
+        _type: Entities.ADMIN._type,
+        _class: Entities.ADMIN._class,
         id: admin.admin_id,
         active: admin.status.toLowerCase() === 'active',
         admin: true,
@@ -112,11 +115,11 @@ export function convertToken(
       source: token,
       assign: {
         _key: `mfa-device:${token.token_id}`,
-        _type: 'mfa_device',
-        _class: 'AccessKey',
+        _type: Entities.MFA_DEVICE._type,
+        _class: Entities.MFA_DEVICE._class,
         id: token.token_id,
-        name: token.token_id,
-        displayName: token.token_id,
+        name: token.token_id || 'name',
+        displayName: token.token_id || 'name',
         serial: token.serial,
         type: token.type,
         factorType: 'token',
@@ -133,11 +136,11 @@ export function convertU2fToken(
       source: token,
       assign: {
         _key: `mfa-device:${token.registration_id}`,
-        _type: 'mfa_device',
-        _class: 'AccessKey',
+        _type: Entities.MFA_DEVICE._type,
+        _class: Entities.MFA_DEVICE._class,
         id: token.registration_id,
-        name: token.registration_id,
-        displayName: token.registration_id,
+        name: token.registration_id || 'name',
+        displayName: token.registration_id || 'name',
         createdOn: getTime(token.date_added),
         factorType: 'u2f',
       },
@@ -153,13 +156,80 @@ export function convertWebAuthnToken(
       source: token,
       assign: {
         _key: `mfa-device:${token.webauthnkey}`,
-        _type: 'mfa_device',
-        _class: 'AccessKey',
+        _type: Entities.MFA_DEVICE._type,
+        _class: Entities.MFA_DEVICE._class,
         id: token.webauthnkey,
-        name: token.credential_name,
+        name: token.credential_name || 'name',
         displayName: `${token.credential_name} ${token.webauthnkey}`,
         createdOn: getTime(token.date_added),
         factorType: 'webauthn',
+      },
+    },
+  });
+}
+
+/**
+ * Devices must validate the data-model property `platform`
+ */
+function getPlatform(duoPlatform: string): string {
+  const dataModelPlatforms = [
+    'darwin',
+    'linux',
+    'unix',
+    'windows',
+    'android',
+    'ios',
+    'embedded',
+  ];
+  for (const dataModelPlatform of dataModelPlatforms) {
+    if (new RegExp(dataModelPlatform, 'i').test(duoPlatform)) {
+      return dataModelPlatform;
+    }
+  }
+  return 'other';
+}
+
+export function convertPhone(
+  phone: DuoPhone,
+): ReturnType<typeof createIntegrationEntity> {
+  return createIntegrationEntity({
+    entityData: {
+      source: phone,
+      assign: {
+        _key: phone.phone_id,
+        _type: Entities.PHONE._type,
+        _class: Entities.PHONE._class,
+        id: phone.phone_id,
+        name: phone.name,
+        category: 'mobile',
+        make: 'UNKNOWN',
+        platform: getPlatform(phone.platform),
+        model: phone.model,
+        encrypted: phone.encrypted === 'Encrypted',
+        serial: 'UNKNOWN',
+      },
+    },
+  });
+}
+
+export function convertIntegration(
+  integration: DuoIntegration,
+): ReturnType<typeof createIntegrationEntity> {
+  (integration as any).secret_key = undefined;
+  const notes = integration.notes;
+  return createIntegrationEntity({
+    entityData: {
+      source: integration,
+      assign: {
+        _key: integration.integration_key,
+        _type: Entities.INTEGRATION._type,
+        _class: Entities.INTEGRATION._class,
+        id: integration.integration_key,
+        name: integration.name,
+        notes:
+          notes !== undefined && notes !== null && !Array.isArray(notes)
+            ? [notes]
+            : notes,
       },
     },
   });